Better Org mode hyperlink syntax highlighting and disabled link conceal
Org mode hyperlinks are now fully syntax highlighted (similar to
Markdown links).

The `org_conceal_links` option has been removed in favour of using
built-in Vim options to enable concealing (with a short section in the
`doc/org.txt` file on how to do this).  This was done because concealed
text makes navigating "hard-wrapped" documents difficult (especially
when the concealed text is as long as most URLs).